+++ app/controllers/flickr.js
@@ -26,16 +26,22 @@ $.init = function() {
 
 	$.retrieveData();
 
-	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#000");
+	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
 
 	if(CONFIG.isChild === true) {
-		$.NavigationBar.showBack();
+		$.NavigationBar.showBack(function(_event) {
+			APP.removeChild();
+		});
 	}
 
 	if(APP.Settings.useSlideMenu) {
-		$.NavigationBar.showMenu();
+		$.NavigationBar.showMenu(function(_event) {
+			APP.toggleMenu();
+		});
 	} else {
-		$.NavigationBar.showSettings();
+		$.NavigationBar.showSettings(function(_event) {
+			APP.openSettings();
+		});
 	}
 };
 
@@ -55,12 +61,13 @@ $.retrieveData = function(_force, _callback) {
 			}
 		},
 		error: function() {
-			Alloy.createWidget("com.chariti.toast", null, {
-				text: "Unable to connect; try again later",
-				duration: 2000
-			});
-
 			APP.closeLoading();
+
+			Alloy.createWidget("com.mcongrove.toast", null, {
+				text: "Unable to connect; try again later",
+				duration: 2000,
+				view: APP.GlobalWrapper
+			});
 
 			if(typeof _callback !== "undefined") {
 				_callback();
@@ -79,7 +86,7 @@ $.handleNsid = function() {
 		cache: CONFIG.cache,
 		callback: $.handleSets,
 		error: function() {
-			Alloy.createWidget("com.chariti.toast", null, {
+			Alloy.createWidget("com.mcongrove.toast", null, {
 				text: "Unable to connect; try again later",
 				duration: 2000
 			});