+++ app/controllers/flickr_album.js
@@ -25,17 +25,20 @@ $.init = function() {
 		cache: CONFIG.cache,
 		callback: $.handleData,
 		error: function() {
-			Alloy.createWidget("com.chariti.toast", null, {
+			Alloy.createWidget("com.mcongrove.toast", null, {
 				text: "Unable to connect; try again later",
-				duration: 2000
+				duration: 2000,
+				view: APP.GlobalWrapper
 			});
 		}
 	});
 
-	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#000");
+	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
 
 	if(APP.Device.isHandheld) {
-		$.NavigationBar.showBack();
+		$.NavigationBar.showBack(function(_event) {
+			APP.removeChild();
+		});
 	}
 };