+++ app/controllers/flickr_photo.js
@@ -14,12 +14,11 @@ var NEXT = null;
 
 var metaVisible = true;
 
-$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#000");
-$.NavigationBar.showBack({
-	callback: function(_event) {
-		APP.removeChild();
-		APP.SlideMenuEngaged = true;
-	}
+$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
+
+$.NavigationBar.showBack(function(_event) {
+	APP.removeChild();
+	APP.SlideMenuEngaged = true;
 });
 
 APP.SlideMenuEngaged = false;