+++ app/controllers/settings_credits.js
@@ -12,12 +12,10 @@ var APP = require("core");
 $.init = function() {
 	APP.log("debug", "settings_credits.init");
 
-	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#000");
+	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
 
-	$.NavigationBar.showBack({
-		callback: function(_event) {
-			APP.removeChild(true);
-		}
+	$.NavigationBar.showBack(function(_event) {
+		APP.removeChild(true);
 	});
 };