+++ app/controllers/text.js
@@ -11,17 +11,23 @@ var CONFIG = arguments[0];
 APP.log("debug", "text | " + JSON.stringify(CONFIG));
 
 $.heading.text = CONFIG.heading;
-$.heading.color = APP.Settings.colors.primary || "#666";
+$.heading.color = APP.Settings.colors.hsb.primary.b > 70 ? "#000" : APP.Settings.colors.primary;
 $.text.text = CONFIG.text;
 
-$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#FFF");
+$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
 
 if(CONFIG.isChild === true) {
-	$.NavigationBar.showBack();
+	$.NavigationBar.showBack(function(_event) {
+		APP.removeChild();
+	});
 }
 
 if(APP.Settings.useSlideMenu) {
-	$.NavigationBar.showMenu();
+	$.NavigationBar.showMenu(function(_event) {
+		APP.toggleMenu();
+	});
 } else {
-	$.NavigationBar.showSettings();
+	$.NavigationBar.showSettings(function(_event) {
+		APP.openSettings();
+	});
 }