+++ app/controllers/web.js
@@ -28,16 +28,29 @@ $.init = function() {
 		$.container.html = CONFIG.html;
 	}
 
-	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#000");
+	$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
 
 	if(CONFIG.isChild === true) {
-		$.NavigationBar.showBack();
+		$.NavigationBar.showBack(function(_event) {
+			APP.removeChild();
+		});
 	}
 
 	if(APP.Settings.useSlideMenu) {
-		$.NavigationBar.showMenu();
+		$.NavigationBar.showMenu(function(_event) {
+			APP.toggleMenu();
+		});
 	} else {
-		$.NavigationBar.showSettings();
+		$.NavigationBar.showSettings(function(_event) {
+			APP.openSettings();
+		});
+	}
+
+	// Move the UI down if iOS7+
+	// NOTE: This is because of problems surrounding a vertical layout on the wrapper
+	//       so we have to bump down the container (47dp normally, 67dp for iOS 7+)
+	if(OS_IOS && APP.Device.versionMajor >= 7) {
+		$.container.top = "67dp";
 	}
 };