+++ app/controllers/youtube_video.js
@@ -14,8 +14,10 @@ $.container.url = CONFIG.url || "";
 $.container.scalesPageToFit = true;
 $.container.willHandleTouches = false;
 
-$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary || "#000");
+$.NavigationBar.setBackgroundColor(APP.Settings.colors.primary);
 
 if(APP.Device.isHandheld) {
-	$.NavigationBar.showBack();
+	$.NavigationBar.showBack(function(_event) {
+		APP.removeChild();
+	});
 }