+++ CHANGELOG.md
@@ -3,6 +3,28 @@ Change Log
 
 This document describes the high-level changes made for each version. For a complete list of changes, please refer to the [ChariTi GitHub repository](http://github.com/mcongrove/ChariTi/).
 
+v 1.2.2 (11-??-2013)
+--------------------
+*	StatusBar styling now based on themes
+*	Added menu header for Settings if menu headers used elsewhere
+*	Better color determinations
+*	Easier-to-use map settings
+*	Fixed missing toolbar on web view
+*	Fixed UI problems for PDF screen
+*	Fixed UI inconsistencies for Settings screen
+*	Fixed file misnaming, misreferences
+*	Widgets are now self-contained
+
+v 1.2.1 (11-18-2013)
+--------------------
+*	Using native toast notifications for Android
+*	Fixed detail navigation for Android
+*	Fixed podcast auto-play duplication
+*	Fixed slide menu headers on Android
+*	Fixed application update process
+*	Fixed over-the-air updates
+*	Improved framework version detection
+
 v 1.2.0 (11-14-2013)
 --------------------
 *	Support for iOS 7